feat(#45): misc enhancements
