+++ webcomponents/core/src/components/slides/deckdeckgo-slide-countdown/deckdeckgo-slide-countdown.scss
@@ -1,0 +1,51 @@
+@import "../deckdeckgo-slide";
+
+div.deckgo-slide {
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+::slotted([slot="title"]), ::slotted([slot="content"]) {
+  text-align: center;
+}
+
+div.deckgo-countdown-container {
+  padding-bottom: var(--slide-countdown-container-padding-bottom, 64px);
+
+  width: inherit;
+  max-width: calc(3 * var(--slide-countdown-digits-width, 16em));
+
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(var(--slide-countdown-digits-width, 16em), 1fr));
+
+  div.time-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    div.figure-container {
+      display: flex;
+      text-align: center;
+
+      div.figure {
+        width: var(--slide-countdown-digit-width, 6em);
+        height: var(--slide-countdown-digit-height);
+        background-color: var(--clock-background-color, #fff);
+        border-radius: var(--slide-countdown-digit-border-radius, 0.5em);
+        box-shadow: var(--slide-countdown-digit-box-shadow, 0 3px 4px 0 rgba(0, 0, 0, .2), inset 2px 4px 0 0 rgba(255, 255, 255, .08));
+
+        &:first-of-type {
+          margin-right: var(--slide-countdown-digit-margin-right, 0.625em);
+        }
+
+        > span {
+          font-size: var(--slide-countdown-digit-font-size, 5em);
+          font-weight: var(--slide-countdown-digit-font-weight);
+          color: var(--slide-countdown-digit-font-color);
+        }
+      }
+    }
+  }
+}
