--- webcomponents/core/src/components/slides/deckgo-slide-countdown/deckdeckgo-slide-countdown.scss
@@ -1,133 +1,0 @@
-@import "../deckdeckgo-slide";
-
-// Custom properties
-// --countdown-base-size : Initial component base size. 
-// --clock-background-color : Clock's background color
-// --clock-foreground-color : Clock's digits' color
-
-:host {
-  font-size: var(--countdown-base-size, 16px);
-}
-
-// Title and content
-::slotted([slot=title]),
-::slotted([slot=content]) {
-  text-align: center;
-}
-
-div.deckgo-slide {
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-// Countdown
-.countdown {
-  display: flex;
-
-  .time-container {
-    margin-right: 2.82em;
-    text-align: center;
-
-    .count-title {
-      display: block;
-      margin-bottom: .94em;
-      margin-top: .94em;
-      font-style: normal;
-      font-variant-ligatures: normal;
-      font-variant-caps: normal;
-      font-variant-numeric: normal;
-      font-variant-east-asian: normal;
-      font-weight: normal;
-      font-stretch: normal;
-      font-size: 0.94em;
-      line-height: normal;
-      color: #1a1a1a;
-      text-transform: uppercase;
-    }
-
-    .figure-container {
-      display: flex;
-
-      .figure {
-        position: relative;
-        float: left;
-        height: 6.87em;
-        width: 6.25em;
-        margin-right: .625em;
-        background-color: var(--clock-background-color, #fff);
-        border-radius: 0.5em;
-        box-shadow: 0 3px 4px 0 rgba(0, 0, 0, .2), inset 2px 4px 0 0 rgba(255, 255, 255, .08);
-
-        &:last-child {
-          margin-right: 0;
-        }
-
-        > span {
-          position: absolute;
-          left: 0;
-          right: 0;
-          margin: auto;
-          font-style: normal;
-          font-variant-ligatures: normal;
-          font-variant-caps: normal;
-          font-variant-numeric: normal;
-          font-variant-east-asian: normal;
-          font-weight: 700;
-          font-stretch: normal;
-          font-size: 5.94em;
-          line-height: 1.15em;
-          font-weight: 700;
-          color: var(--clock-foreground-color, #de4848);
-        }
-
-        .top {
-          z-index: 3;
-          background-color: var(--clock-background-color, #fff);
-          transform-origin: 50% 100%;
-          -webkit-transform-origin: 50% 100%;
-          border-top-left-radius: .625em;
-          border-top-right-radius: .625em;
-          transform: 12.5em;
-          perspective: 12.5em;
-          height: 50%;
-          overflow: hidden;
-          backface-visibility: hidden;
-
-          &:after {
-            content: "";
-            position: absolute;
-            z-index: -1;
-            left: 0;
-            bottom: 0;
-            width: 100%;
-            height: 100%;
-            border-bottom: 1px solid rgba(0, 0, 0, .1);
-          }
-        }
-
-        .bottom {
-          z-index: 1;
-
-          &:before {
-            content: "";
-            position: absolute;
-            display: block;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 50%;
-            background-color: rgba(0, 0, 0, .02);
-          }
-        }
-      }
-    }
-  }
-}
-
-@media screen and (max-width: 786px) {
-  .countdown {
-    flex-direction: column;
-  }
-}
