+++ webcomponents/core/src/index.html
@@ -16,8 +16,11 @@
 
 <deckgo-deck id="slider">
 
-  <deckgo-slide-countdown hours="0" minutes="0" seconds="5">
+  <deckgo-slide-countdown hours="2" minutes="0" seconds="5">
     <h1 slot="title">My presentation will start in</h1>
+    <span slot="hours">Hours</span>
+    <span slot="minutes">Minutes</span>
+    <span slot="seconds">Seconds</span>
   </deckgo-slide-countdown>
 
   <deckgo-slide-title>
